Leet users remove the last protocol from an item
The item still needs at least one protocol to be valid.
This only addresses the case where the user has permission to
apply a single protocol to an item, and that protocol is already
enabled for the item. In this case, if the user removes it, they will no
longer be able to edit the item.