Fix SASS Lint errors
Correct the lint issues flagged in each Gulp SASS build. Note that SASS
Lint doesn't know about `text-decoration-thickness`, so it flags it as
misspelled; I disabled the test in the flagged file, as we know it's not
a misspelling.

Also note that SASS Lint is discontinued, and has not received updates
since 2017. See https://github.com/sasstools/sass-lint

I suggest replacing it with Stylelint. See https://stylelint.io/