655 Fix multiple collections links for a11y (#756)
* 655: Hide duplicate collections link from accessibility tree

This is not a WCAG requirement, but still considered best practice.
See https://accessibleweb.com/question-answer/what-should-i-do-if-i-have-multiple-links-that-go-to-the-same-page/

* 655: Hide icons from accessibility tree

While investigating the `/collections` page, I noticed a flagged a11y
issue around the SVG icons. Icons should be hidden from the
accessibility tree if they are purely presentational, which these icons
all seem to be. I also minified them all to save a few bytes.